Add-edit-expenese

+++ manage-expense.php
@@ -5,12 +5,13 @@ include('includes/dbconnection.php');
 if (strlen($_SESSION['detsuid']==0)) {
   header('location:logout.php');
   } else{
+$userid=$_SESSION['detsuid'];
 
 //code deletion
 if(isset($_GET['delid']))
 {
 $rowid=intval($_GET['delid']);
-$query=mysqli_query($con,"delete from tblexpense where ID='$rowid'");
+$query=mysqli_query($con,"delete from tblexpense where ID='$rowid' && UserId='$userid'");
 if($query){
 echo "<script>alert('Record successfully deleted');</script>";
 echo "<script>window.location.href='manage-expense.php'</script>";
@@ -83,7 +84,6 @@ echo "<script>alert('Something went wrong. Please try again');</script>";
                 </tr>
               </thead>
               <?php
-              $userid=$_SESSION['detsuid'];
 $ret=mysqli_query($con,"select * from tblexpense where UserId='$userid'");
 $cnt=1;
 while ($row=mysqli_fetch_array($ret)) {
@@ -96,7 +96,7 @@ while ($row=mysqli_fetch_array($ret)) {
                   <td><?php  echo $row['ExpenseItem'];?></td>
                   <td><?php  echo $row['ExpenseCost'];?></td>
                   <td><?php  echo $row['ExpenseDate'];?></td>
-                  <td><a href="manage-expense.php?delid=<?php echo $row['ID'];?>">Delete</a>
+                  <td><a href="edit-expense.php?editid=<?php echo $row['ID'];?>">Edit</a> | <a href="manage-expense.php?delid=<?php echo $row['ID'];?>" onclick="return confirm('Do you really want to delete this expense?');">Delete</a>
                 </tr>
                 <?php 
 $cnt=$cnt+1;
